refactor(Tests): Move mocks to separate package
Closes https://github.com/psycholold/TapToUnlock/issues/12
